#module4-task1: tests fixes
