+++ src/components/main/main.test.js
@@ -3,6 +3,8 @@ import MainPage from '../main/main.jsx';
 import renderer from 'react-test-renderer';
 import {BrowserRouter} from "react-router-dom";
 
+jest.mock(`../map/map.jsx`, () => jest.fn().mockReturnValue(null));
+
 const offers = [{
   type: `appartement`,
   price: 140,
@@ -19,7 +21,6 @@ const offers = [{
   rating: 4.0,
   previewPhoto: `img/apartment-02.jpg`,
 }];
-
 
 it(`Main page renders correctly`, () => {
   const tree = renderer
